Separated single tasks logic from project's tasks logic

+++ app/Http/Controllers/CommentController.php
@@ -23,39 +23,43 @@ class CommentController extends Controller
     {
         //
         $validated = $request->validate([
-            'content'=>'required|string'
+            'content' => 'required|string'
         ]);
-        if(!$task){
+        if (!$task) {
             return response()->json([
-                'message'=>'task does not exist'
-            ],404);
+                'message' => 'task does not exist'
+            ], 404);
         }
-        if($task->user() !== $request->user()){
+        if ($task->project()) {
             return response()->json([
-                'message'=>'this action is unauthorized'
-            ],403);
+                'message' => 'this action is not supported'
+            ], 401);
+        }
+        if ($task->user() !== $request->user()) {
+            return response()->json([
+                'message' => 'this action is unauthorized'
+            ], 401);
         }
         $validated['user_id'] = $request->user()->id;
         $comment = $task->comments()->create($validated);
         return response()->json([
-            'status'=>'success',
-            'message'=>'Comment created successfully',
-            'data'=> $comment
+            'status' => 'success',
+            'message' => 'Comment created successfully',
+            'data' => $comment
         ]);
-        
     }
-
+    
     /**
      * Display the specified resource.
-     */
+    */
     public function show(string $id)
     {
         //
     }
-
+    
     /**
      * Update the specified resource in storage.
-     */
+    */
     public function update(Request $request, string $id)
     {
         $validated = $request->validate([
@@ -63,16 +67,21 @@ class CommentController extends Controller
         ]);
         //
         $comment = $request->user()->comments()->find($id);
-        if(!$comment){
+        if (!$comment) {
             return response()->json([
-                'message'=>'comment not found'
-            ],404);
+                'message' => 'comment not found'
+            ], 404);
+        }
+        if ($comment->task->project()) {
+            return response()->json([
+                'message' => 'this action is not supported'
+            ], 403);
         }
         $comment->update($validated);
         return response()->json([
-            'status'=>'success',
-            'message'=>'comment modified successfully',
-            'data'=>$comment
+            'status' => 'success',
+            'message' => 'comment modified successfully',
+            'data' => $comment
         ]);
     }
 
@@ -83,14 +92,19 @@ class CommentController extends Controller
     {
         //
         $comment = $request->user()->comments()->find($id);
-        if(!$comment){
+        if (!$comment) {
             return response()->json([
-                'message'=>'comment not found'
-            ],404);
+                'message' => 'comment not found'
+            ], 404);
+        }
+        if ($comment->task->project()) {
+            return response()->json([
+                'message' => 'this action is not supported'
+            ], 403);
         }
         $comment->delete();
         return response()->json([
-            'message'=>'Comment deleted successfully'
+            'message' => 'Comment deleted successfully'
         ]);
     }
 }
